Delete unused Banner component
As per title
